[Core reorg] Organize source files per area.
