+++ examples/bootstrap_server/CMakeLists.txt
@@ -7,12 +7,14 @@ if(DTLS)
 endif()
 
 include(${CMAKE_CURRENT_LIST_DIR}/../../core/wakaama.cmake)
+include(${CMAKE_CURRENT_LIST_DIR}/../../coap/coap.cmake)
+include(${CMAKE_CURRENT_LIST_DIR}/../../data/data.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/../shared/shared.cmake)
 
 add_definitions(-DLWM2M_BOOTSTRAP_SERVER_MODE)
 add_definitions(${SHARED_DEFINITIONS} ${WAKAAMA_DEFINITIONS})
 
-include_directories (${WAKAAMA_SOURCES_DIR} ${SHARED_INCLUDE_DIRS})
+include_directories (${WAKAAMA_HEADERS_DIR} ${COAP_HEADERS_DIR} ${DATA_HEADERS_DIR} ${WAKAAMA_SOURCES_DIR} ${SHARED_INCLUDE_DIRS})
 
 SET(SOURCES
     ${CMAKE_CURRENT_LIST_DIR}/bootstrap_server.c
@@ -23,7 +25,7 @@ SET(SOURCES
 SET(AUXILIARY_FILES
     ${CMAKE_CURRENT_LIST_DIR}/bootstrap_server.ini)
 
-add_executable(${PROJECT_NAME} ${SOURCES} ${AUXILIARY_FILES} ${WAKAAMA_SOURCES} ${SHARED_SOURCES})
+add_executable(${PROJECT_NAME} ${SOURCES} ${AUXILIARY_FILES} ${WAKAAMA_SOURCES} ${COAP_SOURCES} ${DATA_SOURCES} ${SHARED_SOURCES})
 
 # Add WITH_LOGS to debug variant
 set_property(TARGET ${PROJECT_NAME} APPEND PROPERTY COMPILE_DEFINITIONS $<$<CONFIG:Debug>:WITH_LOGS>)
